create caches for Question/Section/Exam components, may help with #73. Also reorganize files and clean imports
