+++ package.json
@@ -14,7 +14,7 @@
     "build-frontend": "webpack --config webpack.config.js",
     "test": "TS_NODE_FILES=true mocha --recursive -r ts-node/register 'test/**/*.spec.ts'",
     "cypress:open": "CYPRESS_CRASH_REPORTS=0 cypress open",
-    "cypress:run": "CYPRESS_CRASH_REPORTS=0 cypress run --headed --browser chrome",
+    "cypress:run": "CYPRESS_CRASH_REPORTS=0 cypress run --headless --browser chrome",
     "gen-test-exams": "ts-node --files test/test-exams/gen.ts",
     "serve-test-exams": "serve test/test-exams/out/",
     "docs": "typedoc --options typedoc.config.json"
