+++ src/ExamGrader.ts
@@ -75,23 +75,22 @@
  * @module
  */
 
-import { writeFileSync, mkdirSync, existsSync } from 'fs';
+import { writeFileSync, mkdirSync } from 'fs';
 import { TrustedExamSubmission } from './submissions';
-import { Section, Question, Exam, AssignedExam, StudentInfo, RenderMode, AssignedQuestion, AssignedSection, isGradedQuestion } from './exams';
+import { AssignedExam, RenderMode, AssignedQuestion, AssignedSection, isGradedQuestion } from './exams';
 import { QuestionGrader } from './QuestionGrader';
-import { chooseQuestions, chooseSections, CHOOSE_ALL } from './specification';
+import { chooseQuestions, chooseSections, StudentInfo } from './specification';
 import { asMutable, assert, assertFalse, Mutable } from './util';
 import { unparse } from 'papaparse';
 import { createStudentUuid, ExamUtils, writeFrontendJS } from './ExamUtils';
 import { createCompositeSkin, DEFAULT_SKIN } from './skins';
 import del from 'del';
-import { average, chunk, mean, sampleCovariance, standardDeviation, sum } from 'simple-statistics';
-import { GradingAssignmentSpecification } from "./grading/common";
-import { stringify_response } from './response/responses';
-import { renderGradingProgressBar, renderNumBadge, renderPointsProgressBar, renderWideNumBadge } from './ui_components';
+import { average } from 'simple-statistics';
+import { renderGradingProgressBar, renderPointsProgressBar } from './ui_components';
 import { GradedStats } from "./GradedStats";
 import { ExamCurve } from "./ExamCurve";
-import { v4 } from 'uuid';
+import { Exam, Question, Section } from './exam_constructs';
+import { CHOOSE_ALL } from './randomization';
