+++ src/response/util.ts
@@ -1,5 +1,3 @@
-import { MALFORMED_SUBMISSION } from "./common";
-
 export function isNumericArray(x: any) : x is readonly number[] {
     return Array.isArray(x) && x.every(elem => typeof elem === "number");
   }
