+++ test/assigned_exam.spec.ts
@@ -1,24 +1,22 @@
 import 'mocha';
 import { expect } from 'chai';
-import { AssignedExam, Exam } from '../src/exams';
-import { CUSTOMIZE, ExamSpecification, QuestionSpecification, SectionSpecification } from '../src/specification';
-import { MC_Basic } from './question.spec';
+import { AssignedExam } from '../src/exams';
 import { INVALID_IDS, VALID_IDS } from './common.spec';
-import { Section_MC_Basic } from './section.spec';
 import { Exam_MC_Basic } from './exam.spec';
+import { Exam } from '../src/exam_constructs';
 
 
 describe('Exam Specification', () => {
 
   it('Allows Valid Exam IDs', () => {
     VALID_IDS.forEach(
-      id => expect(() => new AssignedExam("uuid", new Exam(Exam_MC_Basic), {name: "student", uniqname: id}, [], false)).not.to.throw()
+      id => expect(() => new AssignedExam("uuid", Exam.create(Exam_MC_Basic), {name: "student", uniqname: id}, [], false)).not.to.throw()
     );
   });
 
   it('Prohibits Invalid Exam IDs', () => {
     INVALID_IDS.forEach(
-      id => expect(() => new AssignedExam("uuid", new Exam(Exam_MC_Basic), {name: "student", uniqname: id}, [], false)).to.throw()
+      id => expect(() => new AssignedExam("uuid", Exam.create(Exam_MC_Basic), {name: "student", uniqname: id}, [], false)).to.throw()
     );
   });
