+++ test/exam.spec.ts
@@ -1,11 +1,9 @@
 import 'mocha';
 import { expect } from 'chai';
-import { ExamSubmission, fillManifest } from '../src/submissions';
-import { Exam, Question, Section } from '../src/exams';
-import { CUSTOMIZE, ExamSpecification, QuestionSpecification, SectionSpecification } from '../src/specification';
-import { MC_Basic } from './question.spec';
+import { CUSTOMIZE, ExamSpecification } from '../src/specification';
 import { INVALID_IDS, VALID_IDS } from './common.spec';
 import { Section_MC_Basic } from './section.spec';
+import { Exam } from '../src/exam_constructs';
 
 export const Exam_MC_Basic : ExamSpecification = {
   id: "exam_id",
@@ -20,13 +18,13 @@ describe('Exam Specification', () => {
 
   it('Allows Valid Exam IDs', () => {
     VALID_IDS.forEach(
-      id => expect(() => new Exam(CUSTOMIZE(Exam_MC_Basic, {id: id}))).not.to.throw()
+      id => expect(() => Exam.create(CUSTOMIZE(Exam_MC_Basic, {id: id}))).not.to.throw()
     );
   });
 
   it('Prohibits Invalid Exam IDs', () => {
     INVALID_IDS.forEach(
-      id => expect(() => new Exam(CUSTOMIZE(Exam_MC_Basic, {id: id}))).to.throw()
+      id => expect(() => Exam.create(CUSTOMIZE(Exam_MC_Basic, {id: id}))).to.throw()
     );
   });
