+++ test/question.spec.ts
@@ -1,9 +1,8 @@
 import 'mocha';
 import { expect } from 'chai';
-import { ExamSubmission, fillManifest } from '../src/submissions';
-import { Question } from '../src/exams';
 import { CUSTOMIZE, QuestionSpecification } from '../src/specification';
 import { INVALID_IDS, VALID_IDS } from './common.spec';
+import { Question } from '../src/exam_constructs';
 
 export const MC_Basic : QuestionSpecification = {
   id: "question_id",
@@ -20,13 +19,13 @@ describe('Question Specification', () => {
 
   it('Allows Valid Question IDs', () => {
     VALID_IDS.forEach(
-      id => expect(() => new Question(CUSTOMIZE(MC_Basic, {id: id}))).not.to.throw()
+      id => expect(() => Question.create(CUSTOMIZE(MC_Basic, {id: id}))).not.to.throw()
     );
   });
 
   it('Prohibits Invalid Question IDs', () => {
     INVALID_IDS.forEach(
-      id => expect(() => new Question(CUSTOMIZE(MC_Basic, {id: id}))).to.throw()
+      id => expect(() => Question.create(CUSTOMIZE(MC_Basic, {id: id}))).to.throw()
     );
   });
